% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{arrange_completed_summary_panels}
\alias{arrange_completed_summary_panels}
\title{Arrange selected completed-summary panels}
\usage{
arrange_completed_summary_panels(panel_grobs, title_text)
}
\arguments{
\item{panel_grobs}{List of ggplot grobs.}

\item{title_text}{Title shown above combined panel.}
}
\description{
Arrange selected completed-summary panels
}
\keyword{internal}
