% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical-metrics.R
\name{calculate_clinical_metrics}
\alias{calculate_clinical_metrics}
\title{Calculate survival statistics for original and imputed data}
\usage{
calculate_clinical_metrics(x, time_points = NULL)
}
\arguments{
\item{x}{bayesian_imputation object}

\item{time_points}{Optional vector of time points for survival probability
estimates. If \code{NULL}, defaults are chosen using \code{time_unit}.}
}
\value{
List with clinical metrics
}
\description{
Calculate survival statistics for original and imputed data
}
\keyword{internal}
