% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calculate_cox_ph_test}
\alias{calculate_cox_ph_test}
\title{Calculate pooled Cox hazard-ratio summary for group comparison}
\usage{
calculate_cox_ph_test(object)
}
\arguments{
\item{object}{bayesian_imputation_groups object}
}
\description{
Calculate pooled Cox hazard-ratio summary for group comparison
}
\keyword{internal}
