% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calculate_survival_probability_diff}
\alias{calculate_survival_probability_diff}
\title{Calculate survival probability difference at specific time point}
\usage{
calculate_survival_probability_diff(object, time_point)
}
\arguments{
\item{object}{bayesian_imputation_groups object}

\item{time_point}{Time point in the same units as the input data}
}
\description{
Calculate survival probability difference at specific time point
}
\keyword{internal}
