% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-comparison.R
\name{combine_model_datasets_safe}
\alias{combine_model_datasets_safe}
\title{Combine completed datasets across compared models}
\usage{
combine_model_datasets_safe(
  object,
  dataset = NULL,
  format = c("wide", "long", "list")
)
}
\arguments{
\item{object}{\code{bird_model_comparison} object.}

\item{dataset}{Dataset index (optional).}

\item{format}{One of \code{"wide"}, \code{"long"}, \code{"list"}.}
}
\description{
Combine completed datasets across compared models
}
\keyword{internal}
