% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-comparison.R
\name{complete.bird_model_comparison}
\alias{complete.bird_model_comparison}
\title{Complete method for model comparison objects}
\usage{
\method{complete}{bird_model_comparison}(
  object,
  dataset = NULL,
  format = c("wide", "long", "list"),
  models = "combined",
  ...
)
}
\arguments{
\item{object}{A \code{bird_model_comparison} object.}

\item{dataset}{Dataset number to extract. \code{NULL} means all (except for wide, where first is used).}

\item{format}{Output format (\code{"wide"}, \code{"long"}, \code{"list"}).}

\item{models}{Either \code{"combined"}, \code{"separate"}, or a specific model name.}

\item{...}{Additional arguments passed to \code{\link[=complete]{complete()}}.}
}
\description{
Complete method for model comparison objects
}
