% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_completed_dataset_output}
\alias{format_completed_dataset_output}
\title{Format completed dataset output columns for user-facing complete()}
\usage{
format_completed_dataset_output(data, time_col = "time")
}
\arguments{
\item{data}{Completed dataset}

\item{time_col}{Name of imputed/current time column in internal objects}
}
\value{
Completed dataset with standardized time columns
}
\description{
Format completed dataset output columns for user-facing complete()
}
\keyword{internal}
