% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDDPsurvival.R
\name{LDDPsurvival.default}
\alias{LDDPsurvival.default}
\title{Default method for LDDPsurvival}
\usage{
\method{LDDPsurvival}{default}(
  formula,
  zpred,
  prior,
  mcmc,
  state,
  status,
  grid,
  data = sys.frame(sys.parent()),
  na.action = na.fail,
  work.dir = NULL
)
}
\arguments{
\item{formula}{survival formula}

\item{zpred}{prediction matrix}

\item{prior}{prior parameters}

\item{mcmc}{MCMC parameters}

\item{state}{initial state}

\item{status}{status vector}

\item{grid}{time grid}

\item{data}{data frame}

\item{na.action}{na.action function}

\item{work.dir}{working directory}
}
\description{
Default method for LDDPsurvival
}
\keyword{internal}
