% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_completed_dataset_summary_models}
\alias{plot_completed_dataset_summary_models}
\title{Plot completed dataset summary for model comparison}
\usage{
plot_completed_dataset_summary_models(
  x,
  dataset_id = NULL,
  alpha = 0.55,
  model_colors = NULL,
  ...
)
}
\arguments{
\item{x}{bird_model_comparison object}

\item{dataset_id}{Dataset index to compare (same index used for all models)}

\item{alpha}{Transparency for overlays}

\item{model_colors}{Optional named/un-named vector of colors for models}

\item{...}{Additional arguments (unused)}
}
\description{
Plot completed dataset summary for model comparison
}
\keyword{internal}
