% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_density_comparison_models}
\alias{plot_density_comparison_models}
\title{Plot density comparison across models (single selected dataset)}
\usage{
plot_density_comparison_models(x, dataset_id = NULL, model_colors = NULL, ...)
}
\arguments{
\item{x}{bird_model_comparison object}

\item{dataset_id}{Dataset index used across all models (NULL for random)}

\item{model_colors}{Optional named/un-named vector of colors for models}

\item{...}{Additional arguments (unused)}
}
\description{
Plot density comparison across models (single selected dataset)
}
\keyword{internal}
