% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_survival_curves}
\alias{plot_survival_curves}
\title{Plot survival curves}
\usage{
plot_survival_curves(
  x,
  n_curves = 10,
  alpha = 0.3,
  show_original = TRUE,
  dataset_id = NULL,
  palette = NULL,
  color = NULL,
  ...
)
}
\arguments{
\item{x}{bayesian_imputation object}

\item{n_curves}{Number of imputed curves to show (default: 10)}

\item{alpha}{Transparency for imputed curves (default: 0.3)}

\item{show_original}{Whether to show original Kaplan-Meier curves (default: TRUE)}
}
\value{
ggplot object
}
\description{
Plot survival curves
}
\keyword{internal}
