% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{plot_survival_curves_separate}
\alias{plot_survival_curves_separate}
\title{Plot survival curves for groups separately}
\usage{
plot_survival_curves_separate(
  x,
  n_curves,
  alpha,
  show_original,
  show_median,
  group_colors,
  ...
)
}
\arguments{
\item{x}{bayesian_imputation_groups object}

\item{n_curves}{Number of curves per group}

\item{alpha}{Transparency}

\item{show_original}{Whether to show original curves}

\item{show_median}{Whether to show dotted median survival guide lines}

\item{group_colors}{Color palette}

\item{...}{Additional arguments}
}
\description{
Plot survival curves for groups separately
}
\keyword{internal}
