% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-comparison.R
\name{print.bird_model_comparison}
\alias{print.bird_model_comparison}
\title{Print method for model comparison objects}
\usage{
\method{print}{bird_model_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{bird_model_comparison} object.}

\item{...}{Unused.}
}
\description{
Print method for model comparison objects
}
