% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_survival_data_metadata}
\alias{strip_survival_data_metadata}
\title{Drop survival_data metadata/class from completed dataset outputs}
\usage{
strip_survival_data_metadata(data)
}
\arguments{
\item{data}{Completed dataset}
}
\value{
Plain data.frame without survival_data attributes
}
\description{
Drop survival_data metadata/class from completed dataset outputs
}
\keyword{internal}
