% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{survival_step_at_grid}
\alias{survival_step_at_grid}
\title{Evaluate a Kaplan-Meier survival step function on a grid}
\usage{
survival_step_at_grid(km, grid)
}
\arguments{
\item{km}{\code{survfit} object.}

\item{grid}{Numeric grid.}
}
\description{
Evaluate a Kaplan-Meier survival step function on a grid
}
\keyword{internal}
