% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{validate_survival_data}
\alias{validate_survival_data}
\title{Validate survival data input}
\usage{
validate_survival_data(data, time_var, status_var, verbose = TRUE)
}
\arguments{
\item{data}{Data frame containing survival data}

\item{time_var}{Name of time variable}

\item{status_var}{Name of status variable}

\item{verbose}{Logical; print status-coding and censoring-pattern messages}
}
\value{
List with validated and processed data
}
\description{
Validate survival data input
}
\keyword{internal}
