% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\name{get_default_priors}
\alias{get_default_priors}
\title{Get default priors for a given distribution}
\usage{
get_default_priors(distribution = "weibull")
}
\arguments{
\item{distribution}{Distribution name ("weibull", "exponential", "lognormal")}
}
\value{
List of default prior parameters
}
\description{
Get default priors for a given distribution
}
